Converted Services to interface and added their implementation
