+++ volunteer-platform/src/main/java/com/example/volunteer_platform/service/EmailService.java
@@ -1,27 +1,16 @@
 package com.example.volunteer_platform.service;
 
-import org.springframework.mail.SimpleMailMessage;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.stereotype.Service;
-import lombok.RequiredArgsConstructor;
+/**
+ * EmailService provides methods to send emails in the system.
+ */
+public interface EmailService {
 
-@Service
-@RequiredArgsConstructor
-public class EmailService {
-    private final JavaMailSender mailSender;
-    
-    public void sendReminderEmail(String to, String subject, String body) {
-        try {
-            SimpleMailMessage message = new SimpleMailMessage();
-            message.setTo(to);
-            message.setSubject(subject);
-            message.setText(body);
-            mailSender.send(message);
-            System.out.println("Email sent successfully to: " + to); // Add this for debug
-        } catch (Exception e) {
-            System.out.println("Failed to send email to: " + to);
-            e.printStackTrace();
-        }
-    }
-
+    /**
+     * Send a reminder email.
+     *
+     * @param to Recipient's email address.
+     * @param subject Subject of the email.
+     * @param body Body of the email.
+     */
+    void sendReminderEmail(String to, String subject, String body);
 }
