+++ volunteer-platform/src/main/java/com/example/volunteer_platform/service/SkillService.java
@@ -1,9 +1,6 @@
 package com.example.volunteer_platform.service;
 
 import com.example.volunteer_platform.model.Skill;
-import com.example.volunteer_platform.repository.SkillRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
@@ -11,20 +8,14 @@ import java.util.Optional;
 /**
  * SkillService provides methods to manage skills in the system.
  */
-@Service
-public class SkillService {
-
-    @Autowired
-    private SkillRepository skillRepository;
+public interface SkillService {
 
     /**
      * Get all skills in the system.
      *
      * @return List of skills.
      */
-    public List<Skill> getAllSkills() {
-        return skillRepository.findAll();
-    }
+    List<Skill> getAllSkills();
 
     /**
      * Find a skill by its name.
@@ -32,9 +23,7 @@ public class SkillService {
      * @param name Name of the skill.
      * @return Optional containing the skill if found.
      */
-    public Optional<Skill> findByName(String name) {
-        return skillRepository.findByName(name);
-    }
+    Optional<Skill> findByName(String name);
 
     /**
      * Find a skill by its ID.
@@ -42,25 +31,19 @@ public class SkillService {
      * @param skillId Skill ID.
      * @return Optional containing the skill if found.
      */
-    public Optional<Skill> findById(Long skillId) {
-        return skillRepository.findById(skillId);
-    }
+    Optional<Skill> findById(Long skillId);
 
     /**
      * Save a skill to the database.
      *
      * @param skill Skill to be saved.
      */
-    public void saveSkill(Skill skill) {
-        skillRepository.save(skill);
-    }
+    void saveSkill(Skill skill);
 
     /**
      * Delete a skill by its ID (admin only).
      *
      * @param skillId Skill ID.
      */
-    public void deleteSkillById(Long skillId) {
-        skillRepository.deleteById(skillId);
-    }
+    void deleteSkillById(Long skillId);
 }
