+++ volunteer-platform/src/main/java/com/example/volunteer_platform/service/TaskService.java
@@ -1,10 +1,6 @@
 package com.example.volunteer_platform.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
 import com.example.volunteer_platform.model.Task;
-import com.example.volunteer_platform.repository.TaskRepository;
 
 import java.util.List;
 import java.util.Optional;
@@ -12,29 +8,21 @@ import java.util.Optional;
 /**
  * TaskService provides methods to manage tasks in the system.
  */
-@Service
-public class TaskService {
-
-    @Autowired
-    private TaskRepository taskRepository;
+public interface TaskService {
 
     /**
      * Get all tasks in the system.
      *
      * @return List of tasks.
      */
-    public List<Task> getAllTasks() {
-        return taskRepository.findAll();
-    }
+    List<Task> getAllTasks();
 
     /**
      * Create a new task associated with an organization.
      *
      * @param task Task to be saved.
      */
-    public void saveTask(Task task) {
-        taskRepository.save(task);
-    }
+    void saveTask(Task task);
 
     /**
      * Find a task by its ID.
@@ -42,18 +30,14 @@ public class TaskService {
      * @param taskId Task ID.
      * @return Optional containing the task if found.
      */
-    public Optional<Task> findById(Long taskId) {
-        return taskRepository.findById(taskId);
-    }
+    Optional<Task> findById(Long taskId);
 
     /**
      * Delete a task by its ID.
      *
      * @param taskId Task ID.
      */
-    public void deleteByTaskId(Long taskId) {
-        taskRepository.deleteById(taskId);
-    }
+    void deleteByTaskId(Long taskId);
 
     /**
      * Search tasks by title, location, or description.
@@ -63,15 +47,5 @@ public class TaskService {
      * @param description Description of the task.
      * @return List of tasks matching the search criteria.
      */
-    public List<Task> searchTasks(String title, String location, String description) {
-        if (title != null) {
-            return taskRepository.findByTitleContaining(title);
-        } else if (location != null) {
-            return taskRepository.findByLocationContaining(location);
-        } else if (description != null) {
-            return taskRepository.findByDescriptionContaining(description);
-        } else {
-            return taskRepository.findAll();
-        }
-    }
+    List<Task> searchTasks(String title, String location, String description);
 }
